+++ src/textarea/index.ts
@@ -1,2 +1,3 @@
 export type {TextareaBridge} from './TextareaBridge';
 export {HiddenTextareaBridge} from './HiddenTextareaBridge'
+export { normalizeTextareaContent } from './utils';